moved unused component and updated records to model
